requirements.txt
Django==4.2.9
djangorestframework==3.14.0
pymongo==4.6.1
djongo==1.3.6
sqlparse==0.2.4
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
django-cors-headers==4.3.1
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
locust==2.20.0
